Add manage page for admins + add process creation for mc server

+++ integration/app.js
@@ -18,6 +18,7 @@ const { Server } = require("socket.io");
 const { serverInfo } = require("./functions");
 const io = new Server(httpServer);
 const PORT = 3003;
+const config = require("./config.json");
 
 process.env.TZ = "Australia/Adelaide";
 
@@ -37,17 +38,22 @@ app.use(session({
 
 // Register web - send it the socket instance, and use router for routes.
 const web = require("./website");
+const { minecraftServer } = require("./DataStorage");
 web.register(io);
 
 app.use("/", web.router);
 
 
 client.on("messageCreate", async(message) => {
+    const cfg = await serverInfo.readConfig();
+    
+    // Only allow admins to run commands.
+    if (!cfg.settings?.admins?.includes(message.author.id)) return;
+
     // This command will create a new message for updating and store it in config.
     if (message.content.startsWith("mc!here")) {
         message.delete();
 
-        const cfg = await serverInfo.readConfig();
         const sentMsg = await serverInfo.sendMessage(message.channel);
 
         // Store current channel id in config.
@@ -57,9 +63,28 @@ client.on("messageCreate", async(message) => {
         };
         
         serverInfo.updateConfig(cfg);
+    } else if (message.content.startsWith("mc!relay")) {
+        // Ensure the parent objects have been made before setting. TODO: Make this into a function.
+        cfg.features = (cfg.features || {});
+        cfg.features.discordChatRelay = (cfg.features.discordChatRelay || {});
+
+        cfg.features.discordChatRelay.channelId = message.channel.id;
+
+        await serverInfo.updateConfig(cfg);
+
+        const reply = await message.reply({
+            "content": "Discord chat relay will now use this channel!"
+        });
+
+        setTimeout(() => {
+            message.delete();
+            reply.delete();
+        }, 3000);
     }
 });
 
+// Initialise the minecraft server object.
+minecraftServer.init(config, io);
 
 httpServer.listen(PORT, () => {
     console.log(`[HTTP] Listening on port ${PORT}`);
